pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "flask>=3.1.0",
    "numpy>=2.2.4",
    "openai>=1.66.3",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "streamlit-extras>=0.6.0",
    "streamlit>=1.43.2",
    "twilio>=9.5.0",
]
